+++ src/methods/submitForm.ts
@@ -1,45 +1,24 @@
 import Promise from 'promise-polyfill';
 import fetchPonyfill from 'fetch-ponyfill';
 import objectAssign from 'object-assign';
+import { SubmissionArgs, SubmissionBody, SubmissionResult } from '../forms';
 import { encode, append } from '../util';
 import { version } from '../../package.json';
 import { Session } from '../session';
 
-export interface Props {
-  id?: string;
-  site?: string;
-  form?: string;
-  data?: FormData | object | undefined;
-  endpoint?: string | undefined;
-  clientName?: string;
-  fetchImpl?: typeof fetch | undefined;
-}
-
-interface SuccessResponse {
-  id: string;
-  data: object;
-}
-
-interface ErrorResponse {
-  errors: Array<{
-    field: string;
-    message: string;
-    code: string | null;
-    properties: object;
-  }>;
-}
-
-type ResponseBody = SuccessResponse | ErrorResponse;
-export type Result = { body: ResponseBody; response: Response };
+const now = (): number => {
+  // @ts-ignore
+  return 1 * new Date();
+};
 
 const serializeBody = (data: FormData | object): FormData | string => {
   if (data instanceof FormData) return data;
   return JSON.stringify(data);
 };
 
-const submissionUrl = (props: Props) => {
-  const { id, site, form } = props;
-  const endpoint = props.endpoint || 'https://api.statickit.com';
+const submissionUrl = (args: SubmissionArgs) => {
+  const { id, site, form } = args;
+  const endpoint = args.endpoint || 'https://api.statickit.com';
 
   if (site && form) {
     return `${endpoint}/j/sites/${site}/forms/${form}/submissions`;
@@ -48,7 +27,7 @@ const submissionUrl = (props: Props) => {
   }
 };
 
-const clientHeader = ({ clientName }: Props) => {
+const clientHeader = ({ clientName }: SubmissionArgs) => {
   const label = `@statickit/core@${version}`;
   if (!clientName) return label;
   return `${clientName} ${label}`;
@@ -56,24 +35,23 @@ const clientHeader = ({ clientName }: Props) => {
 
 export default function submitForm(
   session: Session,
-  props: Props = {}
-): Promise<Result> {
-  if (!props.id && !(props.site && props.form)) {
+  args: SubmissionArgs
+): Promise<SubmissionResult> {
+  if (!args.id && !(args.site && args.form)) {
     throw new Error('`site` and `form` properties are required');
   }
 
-  let fetchImpl = props.fetchImpl || fetchPonyfill({ Promise }).fetch;
-  let url = submissionUrl(props);
-  let data = props.data || {};
-  let sessionWithTime = objectAssign({}, session, {
-    // @ts-ignore
-    submittedAt: 1 * new Date()
+  let fetchImpl = args.fetchImpl || fetchPonyfill({ Promise }).fetch;
+  let url = submissionUrl(args);
+  let data = args.data || {};
+  let sessionWithTime = objectAssign({}, session.data(), {
+    submittedAt: now()
   });
 
   append(data, '_t', encode(sessionWithTime));
 
   let headers: { [key: string]: string } = {
-    'StaticKit-Client': clientHeader(props)
+    'StaticKit-Client': clientHeader(args)
   };
 
   if (!(data instanceof FormData)) {
@@ -88,11 +66,10 @@ export default function submitForm(
   };
 
   return fetchImpl(url, request).then(response => {
-    return response.json().then((body: ResponseBody): {
-      body: ResponseBody;
-      response: Response;
-    } => {
-      return { body, response };
-    });
+    return response.json().then(
+      (body: SubmissionBody): SubmissionResult => {
+        return { body, response };
+      }
+    );
   });
 }